Add files via upload
Stroke disease is very common now days. Stroke can damage the different part of human body. And the therapy is best and quick way to recover from this disease. It is very costly treatment because physiotherapist take huge amount of money for physiotherapy and those stroke effected patient who can’t walk easily, it’s very difficult to him\ her to daily visit the hospital. So we are developing cost effective application which help the stroke patients. Patient can perform exercise for their upper limb their own home by the guidnes of our application. This application not only saves the money of patient even this will save the time also. Patients sit in front of kinect device and the system where application is install. Make sure that upper limb of patient should be clear to kinect device. Session consist of different level. Avatar appears on the screen and guide the patient what to do and how to do. Avatar is basically a graphical statue, behaving like an Expert therapist which is used